Fix the import error about fcrypt on WIndows
Fix of the import error when using Windows because of crypt module not existing. Regarding to the 'no file found of passwords.txt or dictionary.txt' you need to create both text files and put them on the same folder because the zip file of the project does not include them. Comment me back for questions or any other problem.